Fix permission denied error for waifu command
Since the container is running as a random user (26:420), it doesn't
have permission to write the `temp.jpg` file. I could have fixed this by
setting permissions on the file, but since this isn't a scratch
container, I felt it best to just write it it `/tmp` where it would
automatically have permission.